feat: submit practice goal to Airtable from form
