+++ src/pages/PracticeForm.jsx
@@ -1,36 +1,88 @@
 // The id for PracticePlan is tbl3FisNmoYOiIucq.
 
 import React, { useState, useEffect } from 'react';
-import App from './App';
 
 function PracticeForm() {
-  //   async function handleSubmit(e) {
-  //     e.preventDefault();
-  //     const payload = {
-  //       records: [
-  //         {
-  //         fields: {
-  //         goal: goal,
-  //         isCompleted: false,
-  //         },
-  //       },
-  //       ],
-  //     };
+  const [practiceSnippet, setPracticeSnippet] = useState([]);
+  const [goal, setGoal] = useState('');
+  const [newSnippet, setNewSnippet] = useState('');
+  const [isLoading, setIsLoading] = useState(false);
+  const [errorMessage, setErrorMessage] = useState('');
+  const [isSaving, setIsSaving] = useState(false);
 
-  //     const options = {
-  //       method: 'GET',
-  //       headers: {
-  //         'Authorization: token',
-  //       },
-  //       body: JSON.stringify(payload),
-  //     };
+  const url = `https://api.airtable.com/v0/${import.meta.env.VITE_BASE_ID}/${import.meta.env.VITE_TABLE_NAME}`;
+  const token = `Bearer ${import.meta.env.VITE_PAT}`;
 
-  //     try {
-  //       })
-  //     }
-  //     // console.log('Goal Submitted:', goal);
-  //     setGoal('');
-  //   }
+  function handleGoalChange(e) {
+    setGoal(e.target.value);
+    console.log(e.target.value);
+  }
+  //1. function for adding a goal/practice snippet
+
+  async function addSnippet(newSnippet) {
+    const payload = {
+      records: [
+        {
+          fields: {
+            // type:
+            // musicalKey:
+            // metronome:
+            // timeSpent:
+            goal: newSnippet.goal,
+            isCompleted: newSnippet.isCompleted,
+          },
+        },
+      ],
+    };
+
+    const options = {
+      method: 'POST',
+      headers: {
+        Authorization: token,
+        'Content-Type': 'application/json',
+      },
+      body: JSON.stringify(payload),
+    };
+
+    try {
+      setIsSaving(true);
+      const resp = await fetch(url, options);
+
+      if (!resp.ok) {
+        throw new Error('Failed to add new practice snippet');
+      }
+
+      const { records } = await resp.json();
+      const savedSnippet = {
+        id: records[0].id,
+        ...records[0].fields,
+      };
+      console.log(savedSnippet);
+      setPracticeSnippet([...practiceSnippet, savedSnippet]);
+    } catch (error) {
+      console.log(error);
+      setErrorMessage(error.message);
+    } finally {
+      setIsSaving(false);
+    }
+  }
+
+  // async function handleSubmit(e) {
+  //   e.preventDefault();
+  //   setGoal('');
+  // }
+
+  async function handleSubmit(e) {
+    e.preventDefault();
+
+    const newSnippet = {
+      goal,
+      isCompleted: false,
+    };
+
+    await addSnippet(newSnippet);
+    setGoal('');
+  }
 
   return (
     <form onSubmit={handleSubmit}>
